+++ CHANGELOG.md
@@ -7,9 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/trussed-dev/cbor-smol/compare/0.4.1...HEAD
+[Unreleased]: https://github.com/trussed-dev/cbor-smol/compare/0.5.0...HEAD
+
+-
+
+## [0.5.0][] - 2024-10-21
+
+[0.5.0]: https://github.com/trussed-dev/cbor-smol/compare/0.4.1...0.5.0
+
+### Changed
 
 - Mark `Error` as non-exhaustive ([#11](https://github.com/trussed-dev/cbor-smol/issues/11))
+- Add support for multiple `heapless` and `heapless-bytes` versions ([#13](https://github.com/trussed-dev/cbor-smol/pull/13)):
+  - Move existing support for `heapless` 0.7 and `heapless-bytes` 0.3 behind features
+  - Add support for `heapless` 0.8 and `heapless-bytes` 0.4
+  - Remove `cbor_serialize_bytes` and `cbor_serialize_extending_bytes` (use `cbor_serialize_to` instead)
 
 ## [0.4.1][] - 2024-10-08